Moving mains to own package for clarity
- Removed from the udp ones from subsystems
- Made util method for mains (startDispatcher)
