+++ src/Subsystem/ElevatorSubsytem/Elevator.java
@@ -9,8 +9,6 @@ import Messaging.Messages.Events.ElevatorStateEvent;
 import Messaging.Messages.Events.ReceiverBindingEvent;
 import Messaging.Messages.SystemMessage;
 import Messaging.Transceivers.Receivers.Receiver;
-import Messaging.Transceivers.TransceiverDMAFactory;
-import Messaging.Transceivers.TransceiverFactory;
 import Messaging.Transceivers.Transmitters.Transmitter;
 import StatePatternLib.Context;
 import Subsystem.Subsystem;
@@ -118,27 +116,5 @@ public class Elevator extends Context implements Subsystem {
         SystemMessage event = receiver.dequeueMessage();
         System.out.println(event);
         return event;
-    }
-
-    /**
-     * All subsystems combined start procedure (using DMA):
-     * 1. Load numElevators from config
-     * 2. Create transceiver factory
-     * 3. Create and start elevator subsystem.
-     */
-    public static void main(String[] args) {
-        // 1. Configure system from JSON
-        System.out.println("\n****** Configuring System ******\n");
-        Config config = (new Configurator().getConfig());
-        config.printConfig();
-
-        // 2. Create factory
-        TransceiverFactory dmaFactory = new TransceiverDMAFactory();
-
-        // 3. Create and start elevator threads
-        for (int i = 0; i < config.getNumElevators(); ++i) {
-            new Thread(new Elevator(i, dmaFactory.createClientReceiver(i),
-                    dmaFactory.createClientTransmitter())).start();
-        }
     }
 }
